move and re-name preview functions
