+++ plugin/src/steps/preview/index.ts
@@ -8,115 +8,10 @@ import { fetchAndCreateSingleNode } from "~/steps/source-nodes/update-nodes/wp-a
 import { formatLogMessage } from "~/utils/format-log-message"
 import { touchValidNodes } from "../source-nodes/update-nodes/fetch-node-updates"
 
-import type { GatsbyHelpers } from "~/utils/gatsby-types"
 import { IPluginOptions } from "~/models/gatsby-api"
-import { OnPageCreatedCallback } from "../../models/preview"
 
 export const inPreviewMode = (): boolean =>
   !!process.env.ENABLE_GATSBY_REFRESH_ENDPOINT
-
-/**
- * during onCreatePage we want to figure out which node the page is dependant on
-and then store that page in state so we can return info about the page to WordPress
-when the page is updated during Previews.
-We do that by finding the node id on pageContext.id
-Ideally we could detect this without the need for pageContext.id.
-There was an attempt to use store.componentDataDependencies but my implementation 
-was buggy and unreliable. @todo it's worth trying to remove the need for 
-pageContext.id again in the future.
- */
-export const savePreviewNodeIdToPageDependency = (
-  helpers: GatsbyHelpers
-): void => {
-  // if we're not in preview mode we don't want to track this
-  if (!inPreviewMode()) {
-    return
-  }
-
-  const { page, getNode } = helpers
-
-  const nodeThatCreatedThisPage =
-    page.context && page.context.id && getNode(page.context.id)
-
-  if (nodeThatCreatedThisPage) {
-    store.dispatch.previewStore.saveNodePageState({
-      nodeId: nodeThatCreatedThisPage.id,
-      page: {
-        path: page.path,
-        updatedAt: page.updatedAt,
-      },
-    })
-  }
-}
-
-/**
- * during onCreatePage we check if the node this page was created from
- * has been updated and if it has a callback waiting for it
- * if both of those things are true we invoke the callback to
- * respond to the WP instance preview client
- */
-export const onCreatePageRespondToPreviewStatusQuery = async (
-  helpers: GatsbyHelpers
-): Promise<void> => {
-  // if we're not in preview mode we don't want to set this up
-  if (!inPreviewMode()) {
-    return
-  }
-
-  const {
-    nodePageCreatedCallbacks,
-    pagePathToNodeDependencyId,
-  } = store.getState().previewStore
-
-  const { page, getNode } = helpers
-
-  if (
-    !nodePageCreatedCallbacks ||
-    !Object.keys(nodePageCreatedCallbacks).length
-  ) {
-    return
-  }
-
-  const nodeIdThatCreatedThisPage =
-    pagePathToNodeDependencyId?.[page.path]?.nodeId
-
-  if (!nodeIdThatCreatedThisPage) {
-    return
-  }
-
-  const nodePageCreatedCallback =
-    nodeIdThatCreatedThisPage &&
-    nodePageCreatedCallbacks[nodeIdThatCreatedThisPage]
-
-  if (
-    !nodeIdThatCreatedThisPage ||
-    typeof nodePageCreatedCallback !== `function`
-  ) {
-    return
-  }
-
-  const nodeThatCreatedThisPage = getNode(nodeIdThatCreatedThisPage)
-
-  if (!nodeThatCreatedThisPage) {
-    helpers.reporter.warn(
-      formatLogMessage(
-        `There was an attempt to call a Preview onPageCreated callback for node ${nodeIdThatCreatedThisPage}, but no node was found.`
-      )
-    )
-    return
-  }
-
-  await nodePageCreatedCallback({
-    passedNode: nodeThatCreatedThisPage,
-    pageNode: page,
-    context: `onCreatePage Preview callback invocation`,
-    status: `PREVIEW_SUCCESS`,
-  })
-
-  store.dispatch.previewStore.unSubscribeToPagesCreatedFromNodeById({
-    nodeId: nodeIdThatCreatedThisPage,
-  })
-}
 
 type PreviewStatusUnion =
   | `PREVIEW_SUCCESS`
@@ -283,70 +178,5 @@ export const sourcePreviews = async (
     ...webhookBody,
     previewParentId: webhookBody.parentId,
     isPreview: true,
-  })
-}
-
-/**
- * Preview callbacks are usually invoked during onCreatePage in Gatsby Preview
- * so that we can send back the preview status of a created page to WP
- * In the case that no page is created for the node we're previewing, we'll
- * have callbacks hanging around and WP will not know the status of the preview
- * So in onPreExtractQueries (which runs after pages are created), we check which
- * preview callbacks haven't been invoked, and invoke them with a "NO_PAGE_CREATED_FOR_PREVIEWED_NODE" status, which sends that status to WP
- * After invoking all these leftovers, we clear them out from the store so they aren't called again later.
- */
-export const onPreExtractQueriesInvokeLeftoverPreviewCallbacks = async (): Promise<
-  void
-> =>
-  // check for any onCreatePageCallbacks that weren't called during createPages
-  // we need to tell WP that a page wasn't created for the preview
-  invokeAndCleanupLeftoverPreviewCallbacks({
-    status: `NO_PAGE_CREATED_FOR_PREVIEWED_NODE`,
-  })
-
-export const invokeAndCleanupLeftoverPreviewCallbacks = async ({
-  status,
-  context,
-}: {
-  status: string
-  context?: string
-}): Promise<void> => {
-  const state = store.getState()
-
-  const { getNode } = state.gatsbyApi.helpers
-
-  const leftoverCallbacks = state.previewStore.nodePageCreatedCallbacks
-
-  const leftoverCallbacksExist = Object.keys(leftoverCallbacks).length
-
-  if (leftoverCallbacksExist) {
-    await Promise.all(
-      Object.entries(leftoverCallbacks).map(
-        invokeLeftoverPreviewCallback({ getNode, status, context })
-      )
-    )
-
-    // after processing our callbacks, we need to remove them all so they don't get called again in the future
-    store.dispatch.previewStore.clearPreviewCallbacks()
-  }
-}
-
-/**
- * This callback is invoked to send WP the preview status. In this case the status
- * is that we couldn't find a page for the node being previewed
- */
-const invokeLeftoverPreviewCallback = ({ getNode, status, context }) => async ([
-  nodeId,
-  callback,
-]: [string, OnPageCreatedCallback]): Promise<void> => {
-  const passedNode = getNode(nodeId)
-
-  await callback({
-    passedNode,
-    // we pass null as the path because no page was created for this node.
-    // if it had been, this callback would've been removed earlier in the process
-    pageNode: { path: null },
-    status,
-    context,
   })
 }
